Refactor package structure and rename to improve discoverability
- Changed package name from `wsgi-method-override` to `method-override`
- Updated import paths in CI configuration, tests, and middleware
- Removed deprecated `wsgi_method_override` module
- Enhanced README with new installation instructions and usage examples
